# openapi/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/documents/{id}/attachments" : {
      "post" : {
        "tags" : [ "document-attachment-controller" ],
        "operationId" : "upload",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "file" : {
                    "type" : "string",
                    "format" : "binary"
                  }
                },
                "required" : [ "file" ]
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AttachmentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{id}/attachments" : {
      "post" : {
        "tags" : [ "document-attachment-controller" ],
        "operationId" : "upload_1",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "file" : {
                    "type" : "string",
                    "format" : "binary"
                  }
                },
                "required" : [ "file" ]
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AttachmentResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/documents" : {
      "get" : {
        "tags" : [ "document-controller" ],
        "operationId" : "list_6",
        "parameters" : [ {
          "name" : "limit",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "offset",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DocumentListPage"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "document-controller" ],
        "operationId" : "create_9",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateDocumentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DocumentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents" : {
      "get" : {
        "tags" : [ "document-controller" ],
        "operationId" : "list_7",
        "parameters" : [ {
          "name" : "limit",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "offset",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DocumentListPage"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "document-controller" ],
        "operationId" : "create_10",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateDocumentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DocumentResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/attachments/{id}" : {
      "get" : {
        "tags" : [ "attachment-controller" ],
        "operationId" : "get_17",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "Range",
          "in" : "header",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string",
                  "format" : "binary"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "DocumentPermissions" : {
        "type" : "object",
        "properties" : {
          "canEdit" : {
            "type" : "boolean"
          },
          "canPublish" : {
            "type" : "boolean"
          },
          "canManageAccess" : {
            "type" : "boolean"
          },
          "canComment" : {
            "type" : "boolean"
          },
          "canManageAttestations" : {
            "type" : "boolean"
          }
        }
      },
      "DocumentResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "spaceId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "folderId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "title" : {
            "type" : "string"
          },
          "docType" : {
            "type" : "string"
          },
          "body" : {
            "type" : "object",
            "additionalProperties" : { }
          },
          "status" : {
            "type" : "string"
          },
          "currentVersionNo" : {
            "type" : "integer",
            "format" : "int32"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "reliabilityScore" : {
            "type" : "number"
          },
          "reliabilityComputedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "stale" : {
            "type" : "boolean"
          },
          "contentModifiedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "stalenessThresholdDays" : {
            "type" : "integer",
            "format" : "int32"
          },
          "visibility" : {
            "type" : "string"
          },
          "position" : {
            "type" : "integer",
            "format" : "int32"
          },
          "templateId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "templateVersion" : {
            "type" : "integer",
            "format" : "int32"
          },
          "createdBy" : {
            "$ref" : "#/components/schemas/PersonRef"
          },
          "updatedBy" : {
            "$ref" : "#/components/schemas/PersonRef"
          },
          "owner" : {
            "$ref" : "#/components/schemas/PersonRef"
          },
          "tags" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TagRef"
            }
          },
          "permissions" : {
            "$ref" : "#/components/schemas/DocumentPermissions"
          }
        }
      },
      "PersonRef" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "displayName" : {
            "type" : "string"
          },
          "initials" : {
            "type" : "string"
          }
        }
      },
      "TagRef" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "color" : {
            "type" : "string"
          },
          "governed" : {
            "type" : "boolean"
          }
        }
      },
...
      "AttachmentResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "documentId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "filename" : {
            "type" : "string"
          },
          "mediaType" : {
            "type" : "string"
          },
          "sizeBytes" : {
            "type" : "integer",
            "format" : "int64"
          },
          "sha256" : {
            "type" : "string"
          },
          "width" : {
            "type" : "integer",
            "format" : "int32"
          },
          "height" : {
            "type" : "integer",
            "format" : "int32"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "CreateDocumentRequest" : {
        "type" : "object",
        "properties" : {
          "title" : {
            "type" : "string",
            "minLength" : 1
          },
          "body" : {
            "type" : "object",
            "additionalProperties" : { }
          },
          "docType" : {
            "type" : "string"
          },
          "spaceId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "visibility" : {
            "type" : "string"
          },
          "folderId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "templateId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "changeSummary" : {
            "type" : "string"
          }
        },
        "required" : [ "spaceId", "title" ]
      },
...
      "DocumentListPage" : {
        "type" : "object",
        "properties" : {
          "results" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/DocumentSummary"
            }
          },
          "total" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalIsEstimate" : {
            "type" : "boolean"
          },
          "warning" : {
            "type" : "string"
          }
        }
      },
      "DocumentSummary" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "title" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "reliabilityScore" : {
            "type" : "number"
          },
          "stale" : {
            "type" : "boolean"
          },
          "contentModifiedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
    }
  }
}